requirements.txt
Flask>=3.0.0
tensorflow==2.15.0
Pillow>=10.0.0
numpy>=1.24.0,<2.0.0
Flask-CORS>=4.0.0
gunicorn>=21.0.0
opencv-python>=4.8.0
